backend/requirements.txt
# VERITAS Python Requirements
# Install with: pip install -r requirements.txt

groq>=0.4.0
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
pillow>=10.0.0
python-dotenv>=1.0.0
